Only delete page content if the number of connections is zero (#4285)
This PR tries to solve #4253 with a new `_num_connections` counter. On
handshake the counter is increased, on disconnect it is decreased. On
disconnect, after awaiting the reconnect timeout, the page content is
only deleted if the number of connections is zero.

Possible scenarios:

- normal reconnect:
    ```
handshake (num=1) ... disconnect (num=0) ... handshake (num=1) ... after
delay: don't delete
    ```
- strange reconnect with handshake before disconnect:
    ```
handshake (num=1) ... handshake (num=2) ... disconnect (num=1) ... after
delay: don't delete
    ```
- normal disconnect:
    ```
    handshake (num=1) ... disconnect (num=0) ... after delay: delete
    ```

While working on the handshake and disconnect implementation, I noticed
an inconsistency with `app.on_connect` and `app.on_disconnect`: While
connect handlers are called on reconnect, disconnect handlers are not.
In this PR I changed it so that both handlers are called during
reconnection. We should clarify if this is the desired behavior.

+++ nicegui/client.py
@@ -4,6 +4,7 @@ import asyncio
 import inspect
 import time
 import uuid
+from collections import defaultdict
 from contextlib import contextmanager
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Awaitable, Callable, ClassVar, Dict, Iterable, Iterator, List, Optional, Union
@@ -59,8 +60,10 @@ class Client:
         self.environ: Optional[Dict[str, Any]] = None
         self.shared = request is None
         self.on_air = False
-        self._disconnect_task: Optional[asyncio.Task] = None
+        self._num_connections: defaultdict[str, int] = defaultdict(int)
+        self._delete_tasks: Dict[str, asyncio.Task] = {}
         self._deleted = False
+        self._socket_to_document_id: Dict[str, str] = {}
         self.tab_id: Optional[str] = None
 
         self.page = page
@@ -234,30 +237,46 @@ class Client:
         """Add a callback to be invoked when the client disconnects."""
         self.disconnect_handlers.append(handler)
 
-    def handle_handshake(self, next_message_id: Optional[int]) -> None:
+    def handle_handshake(self, socket_id: str, document_id: str, next_message_id: Optional[int]) -> None:
         """Cancel pending disconnect task and invoke connect handlers."""
+        self._socket_to_document_id[socket_id] = document_id
+        self._cancel_delete_task(document_id)
+        self._num_connections[document_id] += 1
         if next_message_id is not None:
             self.outbox.try_rewind(next_message_id)
-        if self._disconnect_task:
-            self._disconnect_task.cancel()
-            self._disconnect_task = None
         storage.request_contextvar.set(self.request)
         for t in self.connect_handlers:
             self.safe_invoke(t)
         for t in core.app._connect_handlers:  # pylint: disable=protected-access
             self.safe_invoke(t)
 
-    def handle_disconnect(self) -> None:
-        """Wait for the browser to reconnect; invoke disconnect handlers if it doesn't."""
-        async def handle_disconnect() -> None:
+    def handle_disconnect(self, socket_id: str) -> None:
+        """Wait for the browser to reconnect; invoke disconnect handlers if it doesn't.
+
+        NOTE:
+        In contrast to connect handlers, disconnect handlers are not called during a reconnect.
+        This behavior should be fixed in version 3.0.
+        """
+        document_id = self._socket_to_document_id.pop(socket_id)
+        self._cancel_delete_task(document_id)
+        self._num_connections[document_id] -= 1
+
+        async def delete_content() -> None:
             await asyncio.sleep(self.page.resolve_reconnect_timeout())
-            for t in self.disconnect_handlers:
-                self.safe_invoke(t)
-            for t in core.app._disconnect_handlers:  # pylint: disable=protected-access
-                self.safe_invoke(t)
-            if not self.shared:
-                self.delete()
-        self._disconnect_task = background_tasks.create(handle_disconnect())
+            if self._num_connections[document_id] == 0:
+                for t in self.disconnect_handlers:
+                    self.safe_invoke(t)
+                for t in core.app._disconnect_handlers:  # pylint: disable=protected-access
+                    self.safe_invoke(t)
+                self._num_connections.pop(document_id)
+                self._delete_tasks.pop(document_id)
+                if not self.shared:
+                    self.delete()
+        self._delete_tasks[document_id] = background_tasks.create(delete_content())
+
+    def _cancel_delete_task(self, document_id: str) -> None:
+        if document_id in self._delete_tasks:
+            self._delete_tasks.pop(document_id).cancel()
 
     def handle_event(self, msg: Dict) -> None:
         """Forward an event to the corresponding element."""
